+++ pol-core/pol/spells.cpp
@@ -51,7 +51,7 @@ Notes
 
 // Magery is repeated at array entry 3, because as of right now, NO spellbook
 // on OSI uses the 301+ spellrange that we can find. 5/30/06 - MuadDib
-u16 spell_scroll_objtype_limits[7][2] = { {0x1F2D,0x1F6C}, {0x2260,0x226F},
+u32 spell_scroll_objtype_limits[7][2] = { {0x1F2D,0x1F6C}, {0x2260,0x226F},
 										  {0x2270,0x227C}, {0x1F2D,0x1F6C},
 										  {0x238D,0x2392}, {0x23A1,0x23A8} , {0x2D51,0x2D60}};
